requirements.txt
fastapi
python-dotenv
uvicorn
psycopg2
sqlalchemy
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
alembic
#fastapi[all] fastapi-mail